update elasticsearch.net to 7.17.2. Make RtqElasticsearchClient compatible with both v6 and v7 elastic: get rid of type name usages when using v7 (which is NOT backward compatible)

+++ Cassandra.DistributedTaskQueue.Monitoring/Storage/Client/TaskSearchResponse.cs
@@ -3,6 +3,6 @@ namespace SkbKontur.Cassandra.DistributedTaskQueue.Monitoring.Storage.Client
     public class TaskSearchResponse
     {
         public string[] Ids { get; set; }
-        public int TotalCount { get; set; }
+        public long TotalCount { get; set; }
     }
 }
